Clarify Bit-Packing doc example positioning (#29)
When reading through the documentation I stumbled upon the Hybrid Bit-Packing section. The section is currently above the deprecated Bit-Packing encoding section, but refers to the example in the latter in a way that presumes the reader has already read it, which left me sightly confused.

Here's an attempt to better point the reader to the other example so that he can easier compare both methods.